requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
gunicorn==23.0.0
numpy==2.1.2
packaging==24.2
pillow==11.0.0
psycopg==3.2.4
psycopg2-binary==2.9.10
setuptools==75.2.0
sqlparse==0.5.1
tzdata==2024.2
wheel==0.44.0
